backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
Flask-Caching
Flask-Mail
psycopg2-binary
SQLAlchemy
PyJWT
Werkzeug
redis
requests
urllib3
numpy
pandas
scikit-learn
scipy
pyTelegramBotAPI
geopy
gunicorn
python-dotenv
python-dateutil
networkx
httpx
pytz
schedule
cloudscraper
fake-useragent
beautifulsoup4
aiohttp
Flask-Migrate
python-slugify==8.0.4
fuzzywuzzy
nltk==3.8.1
joblib==1.3.2
cloudinary
textblob
spacy
cloudinary
Pillow